#11 Details into accordion
